Testing: Emulator and gdb basic functionality with blinky demo.
Smoke test to prove that the qemu emulator and gdb debugger work.
The blinky demo files are known working files used for this purpose.

The Makefile works with BSDmake and GNUmake, and requires that both
arm-none-eabi-gdb and qemu-system-gnuarmeclipse are installed.
Installation instructions and resources will be added in another commit.