Tools/OpenBSD: Add port of STLink V1.1.0 to program STM32 devices.
This is an older version, but it has worked well with my
STM32F4-Discovery board. The current version of STLink is 1.6.1.

To use this port, the Ports System must be installed on your machine.
This is current of OpenBSD 6.6. User-level ports must go under the
directory /usr/ports/mystuff and follow the usual naming conventions
of the Ports System, i.e., this port's directory is
/usr/ports/mystuff/devel/stlink .

To load firmware onto your board:
  $ st-flash --reset write filename.bin 0x08000000

You may need to unplug/replug your usb connector if the programming
was not initially successful.