tools/OpenBSD: Add port of custom qemu for Arm Cortex-M devices.
This is a port for a forked version of qemu 2.8.0 that targets only the
Arm Cortex-M devices in the STM32F{0,1,4} families (and perhaps more).
Only the STM32F4-Discovery board has been tested to work.

See the Makefiles in this repo's "testing" directory for example usage
of qemu, and qemu with gdb debugging.

Note that WXNEEDED needs to be enabled in OpenBSD, since TCG in qemu is
incompatible with W^X protection. This "fixes" a difficult to diagnose
segmentation fault down in the TCG code in qemu.