tools/OpenBSD: Add port of arm-none-eabi-gcc and friends.
This port is the current OpenBSD gcc compiler for embedded Arm.
It is current as of OpenBSD 6.6 and it is saved here for posterity.
This is not one of my ports, as it is not under the "mystuff" directory.

This port is a meta-port and has these sub-projects:
* gcc, Linaro version: 7.4-2019.02, based on gcc 7.4
* binutils: 2.27
* newlib: 2.2.0-1
* gdb: 7.9.1

These are the versions of gcc and friends that will be used for
this project.