testing: Refactor Makefiles under a common scheme.
Makefiles for each directory under "testing" need to only specify
the PROG variable, which will be the same as the assembly file, and
to include the Makefile.inc file. That file has all the make targets.

The top-level "testing" directory has (currently) two make targets:
'build' and 'clean', which runs make with those targets in all
subdirectories specified in the top-level Makefile via the SUBDIR
variable. Each new directory must be appended to SUBDIR.