tools/openbsd: Enable on-chip debugging via OpenOCD and gdb.
Save current (as of OpenBSD 6.6) port of OpenOCD V0.10.0 to this
copy of development ports, for posterity. This is not one of my ports,
as it is not under the "mystuff" directory.

testing: One changed make target name (from gdb to gdb-qemu) and
two new make targets for debugging with OpenOCD and gdb: ocd and gdb-ocd.
The Makefiles have instructions on how to debug with OpenOCD.