testing/002_led: Add a Arm assembly GPIO toggling example
Very thorough example of GPIO manipulation, with all the requiste
register modifying to enable GPIOs to toggle the LEDs.
For each register used, the section of the Reference Manual is referenced.

There are two types of toggling: read/modify/write and atomic set/reset.
The atomic set/reset is due to the STM32F4 having a GPIOx_BSRR register.

Stop showing the graphic in qemu. The console has output that states
whether an LED has changed state.

Use a unified linker script for all examples that is kept with Makefile.inc.