discobsd: Import RetroBSD repo from Feb 17, 2019 Revision 506.
This is a clone of the RetroBSD repository available from
https://github.com/RetroBSD/retrobsd from commit
a0c256c1f0129ee6eb0d00f385e98da9b28d94c1
on Feb 17, 2019. The autobuilder on RetroBSD.org/wiki/autobuild.php
gives this commit the revision of 506.

This commit includes the .gitignore files from the repo, which are
not part of the source archives from the autobuilder.

This is the start of the DiscoBSD project. The name 'DiscoBSD' was
chosen due to the first target of this project is the Arm-based
STM32F4-Discovery board from STMicroelectronics.