resources: Add source archive of RetroBSD Rev 506.
This is a zip archive of the RetroBSD repository at Revision 506
from the autobuilder at http://retrobsd.org/wiki/autobuild.php
and directly relates to the Git commit:
a0c256c1f0129ee6eb0d00f385e98da9b28d94c1

This archive does not include the .gitignore files that a clone of
the repo would include, which is why the Git clone was used
instead of this archive. Saved here for future reference.