tools/openbsd: Add port of devel/mips-elf for MIPS architecture.
This is for 32-bit MIPS architecture.
GCC is version 4.8.1; Binutils is version 2.27.
This toolchain's purpose is to build RetroBSD code on OpenBSD.

A goal of this project is that it becomes, and stays, multiplatform.
All changes to the codebase must work with original MIPS target
and any new ports, such as this DiscoBSD Arm Cortex-M4 port.

Update the ChangeLog to this point.
RetroBSD is a port of 2.11 BSD to MIPS-based PIC32 processors.