build: Remove .MIPS.abiflags section from ELF executables.
The mips-elf toolchain on OpenBSD was adding an additional section
to the ELF executable, in which the elf2aout program was choking
on and unable to handle. This problem was not present on
Ubuntu on Windows Subsystem for Linux.

Bootloaders were having the same problems with overlapping
sections due to the addition of the .MIPS.abiflags section.

The solution I came up with is to remove that section from ever
entering the executable in the first place by adding a /DISCARD/
line to each linker script.

Now, the whole RetroBSD MIPS system builds and runs on OpenBSD.
The VirtualMIPS simulator works, and is able to run a compiled
kernel (not just a pre-made one) and access its filesystem.

Include a screenshot of RetroBSD booting in VirtualMIPS, and link
to it from the repo's README.md file.