knf: Fix warnings for misleading indentation, implicit int.
Haphazard indentation in exec_elf.c. Made it a little more readable.

Misleading indentation in kern_synch.c for panic(). This change
should be the correct intention. In any case, this does not change
the behaviour of the code at all.

Add ints to the few implicit ints of the K&R-style function definitions.

Replace UTF-8 characters with ASCII ones. Fix a typo.