kernel: label_t struct for arm. Used for setjmp/longjmp.
Following the AAPCS standard included in this commit, and precedent
set in the other BSDs, registers r4 to r14 are to be saved into
the label_t jump buffer for setjmp() and longjmp().
Not all 11 registers need to be saved in all cases, but it is better,
and easier, to copy a superset rather than be picky.

A long and an int are the same size on Cortex-M4F, so make it a long
to be consistent with precedent in other BSD systems (notably OpenBSD).

Note that setjmp() and longjmp() have not yet been written for arm.