testing/004_nolibs: libc-free minimal blinky with LL lib.
This testing example is a minimal set of files that will still do
something useful, i.e., blink an LED. This example is directly
based off of 003_cmsis, with all cruft removed and with libc not
linked via the -nostdlib compiler flag.
Also __libc_init_array is not called from startup code.

The Low level library is used for setting up SysTick, GPIO, and Delay.

This is the start to Arm kernel code.