kernel: Tie blinky code into the kernel.
The kernel now fully builds with this commit, although it does
*not* link, as there are many symbols that are required, namely
the u and u0 structs to be defined in the linker script.

All mips assembler code is commented out, to allow the kernel
to compile with arm-none-eabi-gcc.

Add a missing int parameter in gsignal() and a __arm__ define in
sys_pipe.c.

Added -Wno-unused-value and -Wno-uninitialized to temporarily
silence the many warnings they generate with the commented-out code.

Updated Makefile.kconf and Config with those options, as well as
building the vendor-supplied startup code. Regenerating the Makefile
in stm32/f4discovery works well now. It needs the hack to Config
that has two "cpu" config lines: one for mips define and one for the
Arm define lines.

Clock setup code is moved from main.c to machdep.c where it runs
from the MD startup() function. main.c is removed now.

The mips symbols DEVID and TRISE conflicted with vendor-supplied
code, so the mips symbols were renamed.