kernel: The kernel links; the kernel blinky code runs.
Added missing u0, u, u_end symbols to linker script, including
sections of RAM for u0 and u.

Added setjmp()/longjmp() file from newlib-2.2.0.1 for Arm. This
code is untested, but it allows the kernel to link.

The label_t struct has been modified from an array of 11 longs to
and array of 10 longs, to match with the newlib setjmp() code.

Added startup icode code for the handmade /sbin/init process. This
code is also untested, and parts of it are commented out in this
commit. Again, this code is not running, it is merely for linking.

Makefile.kconf is modified to run kconfig with -g switch so that a
debugging kernel is made, which allows gdb to load symbols.