build: Make stm32, arm the default MACHINE{,_ARCH} on DiscoBSD.
Build for Arm by default now. Changing the MACHINE{,_ARCH} variables
on the command line allows the pic32, mips code to be built.

Return VirtualMIPS to targeting the max32 board, as initial testing
for the Arm port is done.

Choose the type of yacc based on operating system via "uname -s".
byacc is used on Linux, yacc is used on everything else.

Removed -msoft-float compiler option for Arm.