report: Structure and outline of report in groff.
Unix resources for text processing have been added to the
resources/unix directory. These are being used to write the report.

The report will be written in groff, using the ms macro package.

Report generation Makefile is made. Automatic regeneration of report
can be achieved via the "entr" program, like so:
  $ ls *.ms | entr -s make
and set up the pdf viewer to reload on file change. In xpdf, the
pdf can be reloaded with the "r" reload command.