report: Add section for Bootstrapping and Linker Script.
Added some more to the System Startup section.

The Linker Script section has a trimmed down kernel linker script
added inside a code box.

A modified version of the example C code for the assembly language
icode routine, which exec's /sbin/init, is included in the
Getting to /sbin/init section of the report inside a code box.
This code is from page 405 of Design and Implementation of 4.3BSD,
and will be referenced as such in a future commit.