build: Turn of -Werror for some games; fix implicit ints.
Add missing implicit ints to some games in src/games.

Turned off -Werror in some Makefiles where there are too many
issues to fix at the moment.

Removed -mips16 CFLAG from program Makefiles, as this is defined
in the target.mk Makefile and is turned on when MACHINE_ARCH=mips
is defined.