uart: Polish uartinit() and uartprobe() code.
Replace PIC32 registers and structs with structs and an array
for STM32 UARTs. Clean up more PIC32 vestiges.
Initialization of all six STM32F4 UARTs is now possible with the
uart[uartnumber] array for configuration data.
Only USART2 currently has a correct configuration, as the others
are place holders for a future commit.

USART2, on pins PA2/PA3, is the system console.

uartprobe() now correctly outputs the initialized uart on the
system bootup console, like so:
  uart2: pins tx=PA2/rx=PA3, af=7, console

Commented out any PIC32 direct register modifying code. This code will
be modified for STM32 in due time.

The t_addr member of a tty struct held the base address of the PIC32
UART registers, and used a non-zero value for error checking if the UART
was initialized or not. The t_addr member now holds the address of the
initialized UART configuration array (or zero, as the case may be) for
the same purpose, but it will not be used for direct access to registers.

kconfig() can now be enabled in init_main.c, as some offending PIC32
spi_setup() code in spi_bus.c has been commented out for the time being.

Add all six UARTs to f4discovery's Config, but do not enable them yet.
USART2 is /dev/tty1, with CONS_MINOR=1.

Replace the MIPS-specific version of ffs() with the one from our libc.
Use ffs() to get the proper gpio pin number for the printout in uartprobe().