sys/stm32: Proper kernel memory map and alignment.
Huge rewrite of the STM32F469NI linker script.
Unused sections (for C++ or Java or heap management) are flensed.
Correct addresses and sizes for the 469 are now used.

Kernel uses STM32 CCM RAM for its data structures and the u and u0
user structures. It starts at 0x1000000 and is 64k. The last 6k is
u0 (3k) and then u (3k). Remaining 58k is available to the kernel.

User RAM will use the standard RAM, which varies between STM32 chips.
The 407 chip has 112k of RAM and the 469 has even more RAM,
but it is best to use the lowest common RAM size for now.
User RAM starts at 0x20000000.

The kernel .text section is in flash and it runs from there.

Kernel and user RAM and flash #defines in machparam.h got a cleaning.
More work to be done, and possibly variables will inherit from the
linker script. Not an easy task for STM32, as there are so many
variations of RAM and flash sizes and would like to support them all.

Alignment issues fixed in setjmp.S when compiled/assembled on Linux.

Properly choose physmem RAM size based on chip ID during cpuidentify().