knf: Replace more UTF-8 with ASCII.
The kernel is clean of UTF-8 characters, except for one of
the keyboard maps in sys/pic32/kbd.c, which is #ifdef'd out
(by USASCII being defined in that file). The current kernel
build has no UTF-8 characters in it.

Some code comments in Russian have been translated to English.

UTF-8 characters have been changed to their equivalent ASCII
characters. Specifically for the license verbiage in the pic32 USB
files, the edits follow the conventions used in other files with
the same license (for example, "(R)" is used for the symbol for
Registered Trademark, as used in sys/pic32/usb_function_hid.c).
The licenses themselves have not been modified.

Many files with UTF-8 characters still exist in the src/ tree.