sys/stm32: Implement uart rx/tx interrupts. write(2) works.
The uart driver is converted from polling to interrupts.

Interrupt priorities are not set in stone, and may need adjustment.
Some uart functionality is not yet implemented or checked (such as
handling a receive buffer overrun).

uartgetc() implemented, and (with uartputc()) uses a uart_inst struct
pointer for all accesses, to be consistent with the rest of the driver.

The receive side of the uart driver is not rigorously tested, but the
transmit side has been tested.
The write(2) system call now works, due to this commit.
Brings back the call to write(2) in init(1).