sys/stm32: Use trap frame SP on return from syscall().
Bug fix for not using stack pointer from u.u_frame after returning
from the call to syscall(). Using the old PSP was a bug.
It is easy now to see how the prologue and the epilogue to syscall()
are (and should be) mirror images of each other.

Now we have a working execve() that doesn't walk over an executable's
argv by a misplaced stack pointer.