sys/stm32: Handle all registers in a syscall, and exec.
The non-AAPCS registers (r4-r11, also named v1-v8) were not saved and
restored during a syscall. They are now pushed and popped in SVC_Handler.

As a consequence, the trapframe and sigcontent structs are expanded
to include registers r4-r11, as well as all code using these structs.

In kernel/exec_subr.c:
exec_clear() is only for clearing the registers, so instead set
default values for lr and psr registers in exec_setupstack().
Expanded the position of the stack pointer (via the trap frame)
after an exec (64 bytes below the executable's argv pointer).