sys/stm32: Enable syscalls that have more than four args.
AAPCS only allows four function parameters to be passed in registers.
Any function parameters greater than four must be passed on the stack.

The fifth and sixth parameters are pushed on the stack, located after the
trapframe (8 and 9 32-bit words, respectively, after the stack pointer).

This commit enables five and six parameter syscalls, which are
sysctl(), pselect(), and select(), and the currently non-implemented
networking syscalls {g,s}etsockopt(), recvfrom(), and sendto().