libc/arm: Enable signal handling in {,sig}{set,long}jmp().
Implemented setjmp(), sigsetjmp(), longjmp(), and siglongjmp().

BSD setjmp() and longjmp() save and restore the signal mask,
while _setjmp() and _longjmp() do not. setjmp() and longjmp() are
functionally equivalent to sigsetjmp() and siglongjmp() when
savesigs is non-zero.

The jump buffer increased to 12 words due to adding savesigs and
the signal mask to the first two spots, respectively, in env[].
For simplicity, jmp_buf[] and sigjmp_buf[] are identical.
_setjmp() clears and _longjmp() skips over those first two words.

setjmp() first saves savesigs (as value 1) and the ten registers
r4-r11, sp (via ip), and lr, then calls sigprocmask() to get the
current sigmask into env[1]. longjmp() does the same in reverse.
longjmp() and _longjmp() are exactly the same except for the
setting of the signal mask in longjmp().