libc/arm: Proper symbols for host/network byte order routines.
Note that this is not an implementation for these routines.

The lack of symbols and function bodies for the htonl,ntohl and
htons,ntohs library routines were a source of bugs that would trample
on the previous or following library routines in the executable.
A nop instruction as function body is sufficient for proper behaviour.