sys/stm32: Clear interrupt masking before execing init.
Due to the differing semantics of the old and new implmentations
of the spl*() functions and interrupt {en,dis}abling, the BASEPRI
register must be explicitly cleared (basically an spl0() call)
before branching to icode() to load and exec init(8).
If not done, init(8) would be stuck in splhigh().
Mask all interrupts before clearing BASEPRI in the set up to user mode.