lib/libc: Favor assembler source files over the C equivalent.
Prefer machine-specific assembler source files for libc routines over
the generic (and potentially more wasteful) C source files.

Precedent for this order of .SUFFIXES for libraries stems from bsd.lib.mk
in 4BSD, if not earlier. Specifically, .S suffix comes before .c suffix.
This is in contrast to the default order of .SUFFIXES in make and gmake.

Although vpath %.S precedes vpath %.c in lib/libc/Makefile, the order of
.SUFFIXES dictates the order (more or less) in which targets are built.
An existing up to date target in ${OBJS} renders the order of vpaths in
the Makefile moot for same-named targets (e.g., bcmp.o).
vpath is a bit of a hack, and should be furloughed.

To compare, src/libc/Makefile handles this situation in another way.
${MACHINE_ARCH} is the last target in ${ALL} to be built, and thus
overwrites the existing file.c-derived file.o with file.S-derived file.o
during the archive extraction loop for the ../libc.a target.
The libc.a archive built and put in /src contains the machine-specific
assembler routine for any conflicting libc routine.