libc/gen: Revive previously disabled libc functions.
Bring these seven libc functions back into the build. No code changes.
As this code is under archaic interface standards, it is throwing
many errors (such as conflicting types) and is breaking the build.
Will be corrected in a forthcoming commit.

Return these seven files to the Makefile.
Alphabetically sort the source and object files in the Makefile.