libc/gen: Fix build errors for seven libc functions.
Fixed the interface errors for the seven libc functions from the
previous commit. Mostly correcting function return types and the
function parameter types, and adding const where required.

Of note is bcopy() and bzero(). Their old return type was implicitly
an int. They are supposed to have a void return type.
ffs(mask) had mask as a long, and now is an int (which are actually
the same on MIPS and Arm architectures).

The build is back.