libc/arm: New implementation of byte order routines for Arm.
New assembly language implementation of 16-bit and 32-bit byte order routines
for Arm Cortex-M that use intrinsic instructions for endian conversions
and zero-extending (in the case of htons() and ntohs()).

Host to network order: 32-bit word htonl(), 16-bit half-word htons()
Network to host order: 32-bit word ntohl(), 16-bit half-word ntohs()

Note that the 32-bit routines are actually the same routine.
Also, the 16-bit routines are the same routine.

The 16-bit routines swap the bytes in each half word, and then
zero-extend the lower half-word to clear out the upper half-word.
The zero-extending is not strictly necessary, since the uint16_t
type should handle that, but is "more correct" and is warranted.