sys: Separate MD and MI types.h (for arch's label_t).
Separate out the machine dependent (MD) struct label_t type from the
rest of the machine independent (MI) types in sys/types.h, and put
it in its own machine/types.h header for each architecture.
Leave pdp11 for now, as it isn't an active architecture in DiscoBSD.

Not wrapping label_t in #ifdef KERNEL as the definition for label_t is
needed for struct user, and some userland programs use it from user.h.