lib: Clean up mips MACHINE_ARCH CFLAGS, ASFLAGS.
Default CFLAGS for all supported architectures, and use ifeq()
to add arch-specific compile flags to CFLAGS (and ASFLAGS).

Compiles in /lib for mips use src/cmd/as as the assembler.
Pass the -x flag to this assembler only when compiling for mips (-Wa,-x)
and add the /lib directory to the search path for executables (-B).
The as assembler uses the -x flag to strip local symbols.

Clean up some comments.