src/arm: Stop zero-filling in Arm userland linker script.
As DiscoBSD is for embedded targets without a memory management unit,
memory is dealt with as fixed and static, not dynamic with "pages".

The userland linker script for Arm had a page-sized (0x10000 bytes)
ALIGN directive, and was zero-filling up to the page and then adding
in the .data and .bss sections afterwards. This made executables huge.

Comment out this ALIGN directive for now. A new linker script to come.

Userland executables now all easily fit in the available SRAM.
The largest program is uucico at around 89kB of SRAM.
The shell is 41kB; emg is 28kB; scm is 72kB; fsck works and is 61kB.
Almost all the executables are smaller on Arm than on pic32 RetroBSD.

Side note:
The kernel currently uses about 92kB Flash and 30kB SRAM (at rest).
The 30kB SRAM *includes* the 6kB for the u0 and u areas (3kB each).