libc/arm: Enable signals, via sigtramp bug fix.
Enable the signal trampoline code to be called, via sendsig()..

Arm Cortex-M processors can only run Thumb-2 code, which has the
specific property that the LSBit of instructions must be set.

Bug missed during the porting of signal code. The fix is to use
the standard ENTRY() define for the sigtramp label, which will
set its .type as %function, and set it as a .thumb_func function.
The compiler and linker handles the rest.