sys/stm32: Handle non-aligned stack frames in syscall().
Fixed an intermittent bug in the handling of arguments 5 and 6 for
syscall(). The issue was that the Arm Cortex-M processor may add a
padding byte to double-word align stack frames if it is needed.

Test the STKALIGN bit (bit 9 in the stacked xPSR) and if it is set then
skip over the padding byte when getting the address for args 5 and 6.

The AAPCS requires that the stack is double-word aligned at a public
interface. The Cortex-M3 and -M4 can automatically handle this by
setting the STKALIGN bit (bit 9) in the Configuration Control Register.
This bit is enabled by default on -M4, but may be cleared by default
on -M3, so set it in SystemInit() just to make sure.

Now all 5 and 6 arg syscalls get the correct arguments.
Programs like cron, hostname, and ps now work properly.