src/cmd: Import arch(1) from OpenBSD. 2-clause BSD-style license.
arch(1) has two names: arch and machine.
machine outputs to stdout the MACHINE define ('stm32' on DiscoBSD).
arch -s outputs to stdout the MACHINE_ARCH define ('arm' on DiscoBSD).

arch (without the -s option) outputs '2.11BSD.arm' to stdout.
This is not preferred, as the dot in '2.11BSD' could cause issues in
shell scripts, since the latter dot is the name separator between the
operating system and the architecture. A future commit may replace
'2.11BSD' with either 'DiscoBSD' or 'RetroBSD', as the case may be.

Added arch(1) and its manpage to the rootfs.manifest* files,
with links for machine(1).

'machine' is analogous with 'uname -m'.
'arch -s' is analogous with 'uname -p'.