cmd/uname: Add -p option to uname(1).
Although the -p option to uname(1) is not a part of POSIX, it is
widely supported and expected across the majority of Unixes.

The -p option will write the machine processor architecture
('arm' on DiscoBSD, 'mips' on RetroBSD) to the standard output.

'uname -p' is analogous to either 'arch -s' or 'machine -a'.

Added the option to the manpage and documented that it is an extension.
Added hostname(1) and machine(1) to the manpage SEE ALSO section.