sys/stm32: Implement reset for halt and reboot.
reboot(8) kills processes and reboots the stm32 microcontroller.
halt(8) does everything reboot(8) does, but then waits for a key press
on the console (or a physical press of the Reset button) before reboot.
Define the HALTREBOOT option for the f469idisco board's kernel.

Using the CMSIS function NVIC_SystemReset() for reset.
Will just spin on a WFI instruction if the reset does not work.

Currently commented out a call to sleep(3) in reboot.c, until the
issue with sleep hanging is figured out.