sys/stm32: SysTick gets a clockframe struct, for hardclock().
Add new clockframe structure, which is the exception stack frame of
a SysTick exception. This stack frame is automatically pushed on the
stack by the hardware. A regular struct trapframe is not appropriate
as it also includes the eight variable registers v1-v8 (not needed in
SysTick) and would require special handling (as is done with syscalls).

Add a SysTick_Handler wrapper in assembly. It determines which stack
the clockframe is on (kernel (MSP) or user (PSP)) and passes a pointer
to that stack to the newly renamed actual SysTick handler, systick().

The call to hardclock() in systick() now has correct parameters
(although frame->cf_psr cannot be used by hardclock(), softclock(), or
gatherstats() until USERMODE() and BASEPRI() are implemented).