sys/stm32: Definitions for USERMODE() and BASEPRI().
USERMODE(psr) returns true if no bits within psr's exception mask are set.
This is the only straightforward way to determine user vs kernel mode, as
the Cortex-M architecture does not have a classical user/kernel mode bit.
It is effective since the xPSR register is in the exception stack frame.

BASEPRI(psr) returns true if the processor currently (meaning, at the
current moment of the call) is not masking any exceptions or interrupts.
BASEPRI() calls the CMSIS function __get_BASEPRI() and tests if the
result returned from the BASEPRI register is zero.
Note that the supplied variable "psr" is not used or needed.
In contrast, the pic32 architecture uses its supplied "ps" value, since
the pic32 stack frame includes the interrupt masking bits (stm32 does not).

All in all, the few instances of BASEPRI() in kern_clock.c are satisfied
with both, equally valid, approaches.

Enable the call to splsoftclock(), prior to softclock(), in kern_clock.c.