sys/stm32: Bring PendSV_Handler() to syscall.c where it belongs.
PendSV_Handler() does the preparation for handling system calls,
and so it should be located in syscall.c with its associates.

Wrap __asm volatile ( ); around the PendSV_Handler assembly language
code. Remove the "bx lr" at the end of the function, as the C
function handles that (verified by disassembly of the executable).