sys/stm32: Add BSP button support to boot single user mode.
Add button support to the STM32F469I-DISCO board support package (BSP)
based on the original HAL-based functions. Port all HAL functions to
use the equivalent LL library functions.

Init the button in machdep.c:startup(), and check if the button is
pressed during boot up, and boot into single user mode if it is.

The button is the blue one on STM32 Discovery and Nucleo boards.

Note that the interrupt-based button code is not in this commit.
The button gpio register state is only checked when the function
BSP_PB_GetState() is called.