libc/arm: Use thumb-only asm in setjmp-family routines.
Convert assembler instructions in the setjmp-family functions to the
equivalent subset of both thumb and thumb2 instruction sets.
For example: 'mov', 'add', 'sub' converted to 'movs', 'adds', 'subs'.
Also, replace the If-Then conditional execution instructions (IT) to
an equivalent compare and label branch idiom.
The 32-bit post-indexing 'str' instruction is replaced with an
equivalent idiom of a 16-bit thumb 'str' followed by an 'adds'.

Note that the remaining 'stmia' and 'ldmia' instructions that operate
on high registers are keeping the setjmp routines thumb2 dirty.
The next commit will clean that up.