build: Enable soft floating point for userland on Arm.
The GCC option -mfpu=fpv4-sp-d16 forces the inclusion of hard
floating point instructions in userland (eventhough it should not).
Remove the option, to be able to use soft floating point libraries.

Also, no need for -mno-thumb-interwork, so get rid of it.

The build now compiles all of userland with soft float, which fixes
the hard faulting on floating point instructions (e.g. vrhadd.u16).