sys/stm32: Fix uart transmission hanging bug. Add comments.
Uart output would intermittently hang until it received some input.
This was due to incorrect handling of when to disable the transmit
interrupt. The proper location is in uartintr() after it is known
that there is a character to send (TXE flag), but before the call
to ttstart()/uartstart().

The call to disable transmit interrupts in uartstart() is redundant,
so remove it.

Terminal handling no longer hangs at all now.

Bring back some comments from 2.11BSD DH11 uart driver, for clarity.