sys/stm32: Add Thumb asm handling of syscall stack frames.
Thumb-only version of stack frame handling code for syscalls in
PendSV_Handler. Select via "#ifdef __thumb2__" by the preprocessor.

The Thumb-only instruction set is very minimal, and most instructions
only work on the "low" registers, i.e., r0-r7. The bulk of the new
Thumb-only code is to dance around stack pushes to and pops from the
low registers in four-register chunks, while also keeping the proper
order of a struct trapframe. There is identical behaviour between
both versions. The Thumb2 version is retained since it is much
smaller and faster than this new Thumb-only version.

Add ".syntax unified" to all inline assembly code, as it is not added
automatically and some instructions require it: e.g., "adds r2, #16".

The file 'syscall.c' now builds with the -mcpu=cortex-m0plus compiler
option and the code is tested and runs on a Cortex-M4.
Will test and run on a Cortex-M0+ once the codebase is Thumb clean.