sys/stm32: Implement kernel {set,long}jmp() for Thumb.
The kernel setjmp() and longjmp() are for context switching, and
are not the same as the libc versions of setjmp() and longjmp().

Conditionally use Thumb2 or Thumb-only assembly for these routines.
As Thumb-only assembler does not use the "high" registers for "ldmia"
and "stmia" and similar instructions, use the idiom of operating with
the "low" registers and then lowering from or hoisting to the "high"
registers. It is slower than Thumb2, so keep both versions.

Still using the large (in .text) block swap code. The Thumb2 version
requires 512 less cycles than the new Thumb-only version for swapping.

Using a bit of a hack with the insertion of a literal pool (by a
.pool directive) in longjmp(), as the large block swap code makes
the label addresses too far away for Thumb-only immediate instructions.