report/log: Remove last remnants of Directed Study project.
The report and project_outputs directories and ChangeLog and README.md
files are the last remaining artifacts of the "Porting the Unix Kernel"
CSC490 Directed Study. As these files have not been updated since the
completion of the Directed Study, remove them from this repository.

Note that these files continue to exist in the CSC490 Directed Study
repository that is archived at https://github.com/chettrick/CSC490