etc/motd: ${MACHINE}-specific motd, with kernel version.
Updated the message of the day to be more like 2.11BSD with the kernel
version, and also a welcome message like 4.4BSD-derived systems.
The welcome message is dependent on ${MACHINE}, with ${MACHINE}-specific
motd files at /etc/etc.${MACHINE}/motd in the source tree.
The particular motd is copied to /etc/motd for distributions.

During boot the rc script will update the kernel version in /etc/motd
if needed, or create a default /etc/motd if it does not exist.

Output the current date before exiting the rc script.

Ignored the generated /etc/motd via .gitignore.