discobsd: Rename /u to /home and enable the file system.
Old versions of BSD used /usr/wd where 'wd' is the initial working
directory of a user, and 'wd' typically is the user's login name.

All modern Unix-like operating systems use /home as the default
location for user home directories, which is better, so do so as well.

The /home file system is now enabled and included in sdcard.img image.