sys/stm32: Make icode() Thumb (ARMv6-M) compatible.
icode() now works for both Thumb2 and Thumb-only.

Thumb does not have a 32-bit load immediate instruction, so it
converts an LDR instruction to a PC-relative load from a literal pool.
The bug that appears in Thumb-only icode() is that the literal pool
does not get copied to user space with the rest of icode().

Force the copying of the literal pool along with the rest of icode()
by placing ".pool" in between the 'icode' and 'sbininit' labels at
the end of any code, but before the arguments for the execv() call.

Force emit the literal pool in longjmp(), so that it doesn't get
mixed up with icode()'s literal pool.

Force the 16-bit version of movs with movs.n, and re-organize the
way icode() is handled: calculate and store the offset first, then
do a register to register adding of the user space base address.
This way works more smoothly with Thumb-only; Thumb2 concurs.

A Thumb-only kernel can now boot and load and exec user space code.