sys/stm32: No HAL_InitTick(); remove *_hal_cortex.[ch] files.
HAL_InitTick() is no longer used. It is kept in the tree, for now,
to aid in porting to other STM32 processors and their HAL libraries.

Remove the files stm32f4xx_hal_cortex.[ch] and replace the calls to
HAL-specific Cortex routines with the CMSIS equivalent routines.

Do not use any NVIC priority grouping. Set it to zero, the default,
and only for Thumb-2 processors.