sys/stm32: Add kernel Config 'board' type #ifdef guards.
Add "#ifdef F469IDISCO" (or whatever the Config 'board' type is)
guards to board-specific source files. This allows the expansion
of source files for new 'board' types that have the same function
API signatures, without "function redefined" compiler errors.

This is simplistic at this stage, and may evolve to a function
pointer-based implementation in the future, if warranted.