sys/stm32: Use blocksize array in BSP_SD_{Read,Write}Blocks().
Instead of using malloc(), use an SD_BLOCK_SIZE array of uint8_t bytes
to hold each block read or written.

SD_BLOCK_SIZE is defined in stm32_adafruit_sd.h as 512 bytes (0x200),
so use that instead of the BlockSize variable.

Add a basic BSP_SD_DeInit() to the BSP API for the Adafruit shield,
to be consistent with the BSP APIs for other boards.