sys/stm32: Disable SPI DMA code. Enable HAL SPI module.
Comment out the DMA code in the HAL SPI library, as it is currently
not being used (and requires a lot of library support to be added).
If needed in the future, it will be easy to bring it back.

Define HAL_SPI_MODULE_ENABLED in stm32f4xx_hal_conf.h to enable the
HAL library SPI module.

Bugfix to remove the reference to stm32f4xx_hal_cortex.h, as it was
missed when that file was removed from the tree.