sys/stm32: Bugfix for kconfig sdio driver config.
Small bugfix for the configuration of the sdio driver in files.kconf
where the 'sdio' driver was erroneously stated as 'sd'.
There is no material change with this now-correct configuration.

Remove the configuration line for 'gpio' for stm32469i_discovery.c
as there actually is not a gpio driver yet.

No change in Makefiles by regenerating.