sys/stm32: Add bare SPI controller driver.
Add basic skeleton of an SPI controller driver and add it to the
character devices as MAJOR 12.
There are potentially eight SPI devices on an STM32F4 microcontroller.

Add a Config kernel configuration for an 'optional' 'spi' controller.
Link 'spi' to the new driver file "spi.c" and STM32 HAL library
file "stm32f4xx_hal_spi.c".

This allows an 'sd' disk to be connected through an 'spi' controller.

Note that the Nucleo and Adafruit SPI library files will be enabled
in files.kconf in a separate commit.